refactor(jni): split the 1,927-line bridge by the domain each export serves
79 `extern "system"` functions in one file. Finding the ink seam meant scrolling past document
lifecycle, every Adjust-sheet setter, and the font registry, and a diff touching one export read as
a diff to the whole boundary — which is the review cost the ~500-line guideline exists to avoid.

The seams were already there in the banner comments; they were just not module boundaries. What the
banners got wrong is worth noting: the one reading "Ink annotation seam" was followed by 500 lines
of zoom, contrast, night mode, crop, viewport, text scale, typeface and column setters before the
first ink function. Splitting by domain rather than by banner is what separates those.

  mod.rs       199   throw/session_mut/dict_ref, the array marshalling, the pagination atomics
  document.rs  475   open, close, render, prefetch, gesture, toc, links, position
  display.rs   485   zoom/fit/crop/contrast/night/viewport, reflow typography, the font registry
  ink.rs       263   the stroke seam and the sidecar store
  lasso.rs     179   selection and the ink clipboard
  text.rs      261   selection, search, and the dictionary
  services.rs   80   the handle-free crate seams (daily, update, opds)

`services.rs` is grouped by a property rather than a subject, deliberately: those four take no
document handle, so they have no aliasing rules to observe and nothing to close, and that is more
useful to know than which crate each one calls.

Every submodule opens with `use super::*` and adds no imports of its own, so there is one import
list to maintain and a helper stays reachable wherever it is used. The pagination atomics moved to
`mod.rs` rather than sitting with the progress exports, because opening a document clears the
cancel flag and installs the sink — `document` needs them too.

This is a move, not a rewrite. `#[unsafe(no_mangle)]` exports by symbol name, not module path, so
the ABI is unchanged, and I checked that rather than asserting it: the 79 symbol names are identical
to master, and so are all 79 function bodies, byte for byte. The Android cross-check
(`cargo check --target aarch64-linux-android --features jni-bridge`) passes, as do 747 host tests
and 166 Kotlin tests.

The IR-7 checker earned its keep here. Reflowing the ink banner into module docs moved the
`IR-7-ALLOW` marker off the line carrying the ADR citation, and the gate failed until the citation
and its exemption were back on the same line.
